fix(dev): intermediate refactor state: consolidating leader
